+++ backend/substrapp/compute_tasks/context.py
@@ -1,3 +1,4 @@
+import os
 from typing import Any
 from typing import Dict
 from typing import List
@@ -8,6 +9,7 @@ from orchestrator import common_pb2
 from orchestrator import computetask_pb2
 from substrapp.compute_tasks.algo import Algo
 from substrapp.compute_tasks.compute_pod import ComputePod
+from substrapp.compute_tasks.directories import SANDBOX_DIR
 from substrapp.compute_tasks.directories import Directories
 from substrapp.orchestrator import get_orchestrator_client
 
@@ -20,11 +22,21 @@ TASK_DATA_FIELD = {
 }
 
 
+class TaskResource(dict):
+    """TaskResource represents a task's input or output.
+
+    By inheriting from dict, we get JSON serialization for free
+    """
+
+    def __init__(self, id: str, value: str):
+        dict.__init__(self, id=id, value=value)
+
+
 class Context:
     """
     Context represents the execution context of a compute task.
 
-    It is read-only, and scoped to a single compute task. It contains all the context data which is useful during the
+    It is scoped to a single compute task. It contains all the context data which is useful during the
     whole lifetime of a compute task: channel name, task key, category, compute plan key, etc...
     """
 
@@ -40,6 +52,7 @@ class Context:
     _directories: Directories
     _algo: Algo
     _has_chainkeys: bool
+    _outputs: Dict[str, str]
 
     def __init__(
         self,
@@ -68,6 +81,7 @@ class Context:
         self._directories = directories
         self._has_chainkeys = has_chainkeys
         self._algo = algo
+        self._outputs = {}
 
     @classmethod
     def from_task(cls, channel_name: str, task: dict):
@@ -171,6 +185,17 @@ class Context:
 
     def get_compute_pod(self, algo_key: str) -> ComputePod:
         return ComputePod(self.compute_plan_key, algo_key)
+
+    def get_output_identifier(self, value: str) -> str:
+        """return the task output identifier from output path"""
+        path = os.path.relpath(value, self.directories.task_dir)
+        return self._outputs[path]
+
+    def set_outputs(self, outputs: List[TaskResource]):
+        """set_outputs should be called with outputs as passed to the algo"""
+        for output in outputs:
+            path = os.path.relpath(output["value"], SANDBOX_DIR)
+            self._outputs[path] = output["id"]
 
 
 def _input_has_kind(input: dict, asset_kind: int, algo: Algo) -> bool: